+++ tests/testthat/test-na.locf.R
@@ -1,6 +1,9 @@
 context("na.locf")
 
-
+test_that("All NA vector throws error",
+          {
+            expect_error(na.locf(c(NA,NA,NA,NA,NA)))
+          })
 
 test_that("Correct results for all options with a modifed tsAirgap dataset (additionalNAs at end)",
           {