+++ tests/testthat/test-na.seasplit.R
@@ -1,6 +1,9 @@
 context("na.seasplit")
 
-
+test_that("All NA vector throws error",
+          {
+            expect_error(na.seasplit(c(NA,NA,NA,NA,NA)))
+          })
 
 test_that("Correct results for all options with a modifed tsAirgap dataset (additionalNAs at end)", {
   skip_on_cran()