+++ tests/testthat/test-na_interpolation.R
@@ -1,6 +1,9 @@
 context("na_interpolation")
 
-
+test_that("All NA vector throws error",
+          {
+            expect_error(na_interpolation(c(NA,NA,NA,NA,NA)))
+          })
 
 test_that("Correct results for all options with a modifed tsAirgap dataset (additionalNAs at end)",
           {