+++ tests/testthat/test-na_ma.R
@@ -1,5 +1,9 @@
 context("na_ma")
 
+test_that("All NA vector throws error",
+          {
+            expect_error(na_ma(c(NA,NA,NA,NA,NA)))
+          })
 
 test_that("Correct results for all options with a modifed tsAirgap dataset (additionalNAs at end)",
           {