+++ tests/testthat/test-na_random.R
@@ -1,5 +1,10 @@
 context("na_random")
 
+test_that("All NA vector throws error",
+          {
+            expect_error(na_random(c(NA,NA,NA,NA,NA)))
+          })
+
 test_that("Wrong input",
           {
             x <- data.frame(tsAirgap)