+++ tests/testthat/test-na_replace.R
@@ -1,5 +1,10 @@
-context("na.replace")
+context("na_replace")
 
+
+test_that("All NA vector throws no error",
+          {
+            expect_that(sum(na_replace(c(NA,NA,NA,NA,NA)), fill = 2.0), is_identical_to(10.0))
+          })
 
 test_that("Correct results for all options with a modifed tsAirgap dataset (additionalNAs at end)",
           {
@@ -10,8 +15,8 @@ test_that("Correct results for all options with a modifed tsAirgap dataset (addi
             #Using rounded version in order to avoid writing down all decimals
             x <- tsAirgap
             x[135:144] <- NA
-            expect_that( round(mean(na.replace( x, fill = -1)), digits = 1),  is_identical_to(223.6) )
-            expect_that( round(mean(na.replace( x, fill = 200 )), digits = 1),  is_identical_to(254.3) )
+            expect_that( round(mean(na_replace( x, fill = -1)), digits = 1),  is_identical_to(223.6) )
+            expect_that( round(mean(na_replace( x, fill = 200 )), digits = 1),  is_identical_to(254.3) )
             })
 
 test_that("Correct results for all options with a modifed tsAirgap dataset (additionalNAs at start)",
@@ -23,8 +28,8 @@ test_that("Correct results for all options with a modifed tsAirgap dataset (addi
             #Using rounded version in order to avoid writing down all decimals
             x <- tsAirgap
             x[1:5] <- NA
-            expect_that( round(mean(na.replace( x, fill = -1)), digits = 1),  is_identical_to(251.0) )
-            expect_that( round(mean(na.replace( x, fill = 200)), digits = 1),  is_identical_to(274.8) )
+            expect_that( round(mean(na_replace( x, fill = -1)), digits = 1),  is_identical_to(251.0) )
+            expect_that( round(mean(na_replace( x, fill = 200)), digits = 1),  is_identical_to(274.8) )
             })
 
 
@@ -36,25 +41,25 @@ test_that("Correct results for all options with the tsAirgap dataset",
             #In order to avoid writing down the complete resulting vector
             #Using rounded version in order to avoid writing down all decimals
             x <- tsAirgap
-            expect_that( round(mean(na.replace( x, fill = -1)), digits = 1),  is_identical_to(254.5) )
-            expect_that( round(mean(na.replace( x, fill = 200 )), digits = 1),  is_identical_to(272.6) )
+            expect_that( round(mean(na_replace( x, fill = -1)), digits = 1),  is_identical_to(254.5) )
+            expect_that( round(mean(na_replace( x, fill = 200 )), digits = 1),  is_identical_to(272.6) )
             })
 
 test_that("Imputation works for data.frame",
           {
             #Checking if NAs remain in data.frame
             x <- data.frame(tsAirgap, tsAirgap, tsAirgapComplete)
-            expect_that(anyNA(na.replace(x, fill = -1 )), is_false())
-            expect_that(anyNA(na.replace(x, fill = 200)), is_false())
+            expect_that(anyNA(na_replace(x, fill = -1 )), is_false())
+            expect_that(anyNA(na_replace(x, fill = 200)), is_false())
               })
 
 test_that("Test NA at beginning",
           {
             x <- tsAirgap
             x[1:2] <- NA
-            expect_that(anyNA(na.replace(x, fill = -5)), is_false())
-            expect_that(anyNA(na.replace(x, fill = 1000)), is_false())
-            expect_that(anyNA(na.replace(x)), is_false())
+            expect_that(anyNA(na_replace(x, fill = -5)), is_false())
+            expect_that(anyNA(na_replace(x, fill = 1000)), is_false())
+            expect_that(anyNA(na_replace(x)), is_false())
             
           })
 
@@ -62,9 +67,9 @@ test_that("Test NA at end",
           {
             x <- tsAirgap
             x[143:144] <- NA
-            expect_that(anyNA(na.replace(x, fill = -5)), is_false())
-            expect_that(anyNA(na.replace(x, fill = 1000)), is_false())
-            expect_that(anyNA(na.replace(x)), is_false())
+            expect_that(anyNA(na_replace(x, fill = -5)), is_false())
+            expect_that(anyNA(na_replace(x, fill = 1000)), is_false())
+            expect_that(anyNA(na_replace(x)), is_false())
             
           })
 
@@ -72,9 +77,9 @@ test_that("Multiple NAs in a row",
           {
             x <- tsAirgap
             x[40:80] <- NA
-            expect_that(anyNA(na.replace(x, fill = -5)), is_false())
-            expect_that(anyNA(na.replace(x, fill = 1000)), is_false())
-            expect_that(anyNA(na.replace(x)), is_false())
+            expect_that(anyNA(na_replace(x, fill = -5)), is_false())
+            expect_that(anyNA(na_replace(x, fill = 1000)), is_false())
+            expect_that(anyNA(na_replace(x)), is_false())
             
           })
 
@@ -82,15 +87,15 @@ test_that("Over 90% NAs",
           {
             x <- tsAirgap
             x[10:140] <- NA
-            expect_that(anyNA(na.replace(x, fill = -5)), is_false())
-            expect_that(anyNA(na.replace(x, fill = 1000)), is_false())
-            expect_that(anyNA(na.replace(x)), is_false())
+            expect_that(anyNA(na_replace(x, fill = -5)), is_false())
+            expect_that(anyNA(na_replace(x, fill = 1000)), is_false())
+            expect_that(anyNA(na_replace(x)), is_false())
             
           })
 
 test_that("Handling for no NAs",
           {
             x <- tsAirgapComplete
-            expect_that(anyNA(na.replace(x)), is_false())
+            expect_that(anyNA(na_replace(x)), is_false())
             
           })